Create a_Water_Park.c
Chef decides to go to the water park to play. To enter the water slide, a person must have a weight of at most W Kg and a height of at least H cm.
Chef weighs 60 kg and his height is 130 cm.
Is Chef allowed to enter the water slide?