Create Chocolate_Distribution.c
Christmas is the season of giving. As a part of celebrating the New Year, you decided to give away chocolates to the poor children in your neighbourhood. Some of them even asked for 2 chocolates and being the kind soul that you are, you always gave them if you could. They never asked for more than 2 though because they are not greedy. Now, you have forgotten exactly how many children you gave chocolates to. However, you remember that you started with N chocolates and distributed all of them by the end.
Find the minimum and maximum number of possible children that you gave chocolates to.